use cobra framework for cli
now budg will take command line arguments to calculate the budget table.
