+++ go.mod
@@ -1,5 +1,13 @@
-module test/budgo
+module github.com/kybouw/budgo
 
 go 1.21.5
 
-require github.com/BurntSushi/toml v1.3.2
+require (
+	github.com/BurntSushi/toml v1.3.2
+	github.com/spf13/cobra v1.3.0
+)
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
